requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
pyyaml
tabulate
lxml
